+++ src/main/java/com/nikolaev/data/exceptions/LoadFailExceptiojn.java
@@ -1,0 +1,11 @@
+package com.nikolaev.data.exceptions;
+
+public class LoadFailExceptiojn extends Exception {
+
+	private static final long serialVersionUID = 3616308636482710195L;
+
+	public LoadFailExceptiojn(String message) {
+		super(message);
+	}
+
+}
