--- src/main/java/data/exceptions/LoadFailExceptiojn.java
@@ -1,11 +1,0 @@
-package data.exceptions;
-
-public class LoadFailExceptiojn extends Exception {
-
-  private static final long serialVersionUID = 3616308636482710195L;
-  public LoadFailExceptiojn(String message) {
-    super(message);
-  }
-  
-
-}
