--- src/main/java/data/idea/Worker.java
@@ -1,8 +1,0 @@
-package data.idea;
-
-
-public class Worker {
-
-  
-
-}
